+++ fiddles/d3/fiddle-0026-Nvd3VerticalBarModel/app.js
@@ -1,5 +1,0 @@
-(function (app, $, undefined) {
-    "use strict";
-    d3.select("p#target")
-        .text("Hello world!");
-})(window.app = window.app || {}, d3)